requirements.txt
python-dotenv
litellm
pyobjc-framework-Quartz
